requirements.txt
chalice~=1.26.6
Faker~=13.3.2
python-dateutil~=2.8.2
